integration of screen manager and xml user file creation
Combined the changes in screens-ver folder with the changes in the main folder

Deleted files are already in the main folder
