--- BisPrend-screens-ver/font-guide.txt
@@ -1,4 +1,0 @@
-FONT GUIDE
-
-Default Font -
-Logo Font - Mont Heavy Demo
