+++ BisPrend/bisprend.kv
@@ -1,20 +1,27 @@
 #:kivy 2.0.0
 
-<BisprendEngine>:
-    player: username
+WindowManager:
+    RegWindow:
+    WelcomeWindow:
+
+<RegWindow>:
+    name: "Reg"
+    username: username #object
+
     FloatLayout:
         orientation: 'vertical'
         canvas:
             Rectangle:
-                size: root.width, root.height
-                source: 'bg-1.png'
+                pos: self.pos
+                size: self.size
+                source: "bg-1.png"
 
         Label:
-            font_size: 85
+            font_size: 100
             font_name: "Mont"
-            center_x: root.width / 2
-            top: root.top - 20
             text: "Pangalan"
+            color: 0.12,0.12,0.12
+            pos_hint: {'center_y': 0.75, 'center_x': 0.5}
             color: 0.12,0.12,0.12
 
         TextInput:
@@ -22,26 +29,36 @@
             id: username
             font_size: 30
             font_name: "Mont"
-            size_hint: (4.5,0.5)
-            center_x: root.width / 2
-            top: root.top - 150
+            size_hint: (0.4, 0.15)
+            pos_hint: {'center_y': 0.55, 'center_x': 0.5}
 
 
         Button:
+            text: "Ok"
             background_color: 0,0,0
-            text: "Confirm"
             color: 1,1,1
             font_size: 30
             font_name: "Mont"
-            size_hint: (4.5,0.5)
-            center_x: root.width / 2
-            top: root.top - 210
-            text: "Ok"
-            color: 1,1,1
-            on_release: root.btn()
+            size_hint: (0.2,0.1)
+            pos_hint: {'center_y': 0.35, 'center_x': 0.5}
+            on_release:
+                root.registerUser()
+                app.root.current = "Welcome"
+                root.manager.transition.direction = "left"
+                
+<WelcomeWindow>:
+    name: "Welcome"
+    FloatLayout:
+        canvas:
+            Rectangle:
+                size: root.width, root.height
+                source: "default-bg.png"
 
-<BisprendPopup>:
-    Label:
-        text: "Welcome!"
-        size_hint: 0.6, 0.8
-        pos_hint: {"x": 0.2, "top": 1}
+        Label:
+            text: "Welcome!" 
+            font_name: "Mont"
+            font_size: 40
+            pos_hint: {'center_y': 0.5, 'center_x': 0.5}
+            #center_x: root.width / 2
+            #top: root.top - 20
+            color: 0.12,0.12,0.12
