--- BisPrend/items.xml
@@ -1,1 +1,0 @@
-<data><items><datum name="name">Kiev Cangs</datum><datum name="progress">0</datum></items></data>
